Add TeamCityOutput via --setup script [BC break]
The output handler is a deliberate step back: it is very simple as it just prints the results without grouping them in any way. But, on a might brighter note, this approach works with pretty much any version of Nette Tester :tada: (a more intelligent output handler implementation waits for the resolution of nette/tester#345)

Breaking change: a user-defined setup script, previously configured via --setup entry in "Tester options", must now be provided in a separate "User setup script" option in the Run Configuration dialog, otherwise tests and/or the plugin might not work as expected.